+++ core/auth/auth.go
@@ -24,20 +24,20 @@ import (
 	"sync"
 	"time"
 
-	"github.com/golang-jwt/jwt"
+	"github.com/golang-jwt/jwt/v5"
 	"github.com/google/uuid"
-	"github.com/rokwire/core-auth-library-go/v3/authorization"
-	"github.com/rokwire/core-auth-library-go/v3/authservice"
-	"github.com/rokwire/core-auth-library-go/v3/authutils"
-	"github.com/rokwire/core-auth-library-go/v3/keys"
-	"github.com/rokwire/core-auth-library-go/v3/sigauth"
-	"github.com/rokwire/core-auth-library-go/v3/tokenauth"
+	"github.com/rokwire/rokwire-building-block-sdk-go/services/core/auth"
+	"github.com/rokwire/rokwire-building-block-sdk-go/services/core/auth/authorization"
+	"github.com/rokwire/rokwire-building-block-sdk-go/services/core/auth/keys"
+	"github.com/rokwire/rokwire-building-block-sdk-go/services/core/auth/sigauth"
+	"github.com/rokwire/rokwire-building-block-sdk-go/services/core/auth/tokenauth"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/rokwireutils"
 	"golang.org/x/sync/syncmap"
 	"gopkg.in/go-playground/validator.v9"
 
-	"github.com/rokwire/logging-library-go/v2/errors"
-	"github.com/rokwire/logging-library-go/v2/logs"
-	"github.com/rokwire/logging-library-go/v2/logutils"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/errors"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/logging/logs"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/logging/logutils"
 )
 
 const (
@@ -97,7 +97,7 @@ type Auth struct {
 
 	authPrivKey *keys.PrivKey
 
-	ServiceRegManager *authservice.ServiceRegManager
+	ServiceRegManager *auth.ServiceRegManager
 	SignatureAuth     *sigauth.SignatureAuth
 
 	serviceID   string
@@ -127,7 +127,7 @@ type Auth struct {
 }
 
 // NewAuth creates a new auth instance
-func NewAuth(serviceID string, host string, authPrivKey *keys.PrivKey, authService *authservice.AuthService, storage Storage, emailer Emailer, minTokenExp *int64,
+func NewAuth(serviceID string, host string, authPrivKey *keys.PrivKey, authService *auth.Service, storage Storage, emailer Emailer, minTokenExp *int64,
 	maxTokenExp *int64, deleteMembershipsPeriod *int64, supportLegacySigs bool, twilioAccountSID string, twilioToken string, twilioServiceSID string, profileBB ProfileBuildingBlock,
 	smtpHost string, smtpPortNum int, smtpUser string, smtpPassword string, smtpFrom string, logger *logs.Logger, version string) (*Auth, error) {
 	if minTokenExp == nil {
@@ -159,67 +159,70 @@ func NewAuth(serviceID string, host string, authPrivKey *keys.PrivKey, authServi
 		deletePeriod = *deleteMembershipsPeriod
 	}
 
-	auth := &Auth{storage: storage, emailer: emailer, logger: logger, authTypes: authTypes, externalAuthTypes: externalAuthTypes, anonymousAuthTypes: anonymousAuthTypes,
+	authStruct := &Auth{storage: storage, emailer: emailer, logger: logger, authTypes: authTypes, externalAuthTypes: externalAuthTypes, anonymousAuthTypes: anonymousAuthTypes,
 		serviceAuthTypes: serviceAuthTypes, mfaTypes: mfaTypes, authPrivKey: authPrivKey, ServiceRegManager: nil, serviceID: serviceID, host: host, minTokenExp: *minTokenExp,
 		maxTokenExp: *maxTokenExp, profileBB: profileBB, cachedIdentityProviders: cachedIdentityProviders, identityProvidersLock: identityProvidersLock, apiKeys: apiKeys,
 		apiKeysLock: apiKeysLock, deleteSessionsTimerDone: deleteSessionsTimerDone, deleteMembershipsTimerDone: deleteMembershipsTimerDone, deleteMembershipsPeriod: deletePeriod, version: version}
 
-	err := auth.storeCoreRegs()
+	err := authStruct.storeCoreRegs()
 	if err != nil {
 		return nil, errors.WrapErrorAction(logutils.ActionSave, model.TypeServiceReg, nil, err)
 	}
-	auth.storeCoreServiceAccount()
+	authStruct.storeCoreServiceAccount()
 
 	serviceRegLoader := NewLocalServiceRegLoader(storage)
 
 	// Instantiate a ServiceRegManager to manage the service registration data loaded by serviceRegLoader
-	serviceRegManager, err := authservice.NewServiceRegManager(authService, serviceRegLoader, true)
+	serviceRegManager, err := auth.NewServiceRegManager(authService, serviceRegLoader, true)
 	if err != nil {
 		return nil, errors.WrapErrorAction(logutils.ActionInitialize, "service reg manager", nil, err)
 	}
 
-	auth.ServiceRegManager = serviceRegManager
+	authStruct.ServiceRegManager = serviceRegManager
 
 	signatureAuth, err := sigauth.NewSignatureAuth(authPrivKey, serviceRegManager, true, supportLegacySigs)
 	if err != nil {
 		return nil, errors.WrapErrorAction(logutils.ActionInitialize, "signature auth", nil, err)
 	}
 
-	auth.SignatureAuth = signatureAuth
+	authStruct.SignatureAuth = signatureAuth
 
 	// auth types
-	initUsernameAuth(auth)
-	initEmailAuth(auth)
-	initPhoneAuth(auth, twilioAccountSID, twilioToken, twilioServiceSID)
-	initFirebaseAuth(auth)
-	initAnonymousAuth(auth)
-	initSignatureAuth(auth)
+	initUsernameAuth(authStruct)
+	initEmailAuth(authStruct)
+	initPhoneAuth(authStruct, twilioAccountSID, twilioToken, twilioServiceSID)
+	initFirebaseAuth(authStruct)
+	initAnonymousAuth(authStruct)
+	initSignatureAuth(authStruct)
 
 	// external auth types
-	initOidcAuth(auth)
-	initSamlAuth(auth)
+	initOidcAuth(authStruct)
+	initSamlAuth(authStruct)
 
 	// service auth types
-	initStaticTokenServiceAuth(auth)
-	initSignatureServiceAuth(auth)
+	initStaticTokenServiceAuth(authStruct)
+	initSignatureServiceAuth(authStruct)
 
 	// mfa types
-	initTotpMfa(auth)
-	initEmailMfa(auth)
-	initPhoneMfa(auth)
-	initRecoveryMfa(auth)
+	initTotpMfa(authStruct)
+	initEmailMfa(authStruct)
+	initPhoneMfa(authStruct)
+	initRecoveryMfa(authStruct)
 
-	err = auth.cacheIdentityProviders()
+	err = authStruct.cacheIdentityProviders()
 	if err != nil {
 		logger.Warnf("NewAuth() failed to cache identity providers: %v", err)
 	}
 
-	err = auth.cacheAPIKeys()
+	err = authStruct.cacheAPIKeys()
 	if err != nil {
 		logger.Warnf("NewAuth() failed to cache api keys: %v", err)
 	}
 
-	return auth, nil
+	// Encode audience claim as single string rather than array if only one item
+	jwt.MarshalSingleStringAsArray = false
+
+	return authStruct, nil
 
 }
 
@@ -2242,14 +2245,14 @@ func (a *Auth) constructServiceAccount(accountID string, name string, appID stri
 	}
 
 	var application *model.Application
-	if appID != authutils.AllApps {
+	if appID != rokwireutils.AllApps {
 		application, err = a.storage.FindApplication(nil, appID)
 		if err != nil || application == nil {
 			return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeApplication, nil, err)
 		}
 	}
 	var organization *model.Organization
-	if orgID != authutils.AllOrgs {
+	if orgID != rokwireutils.AllOrgs {
 		organization, err = a.storage.FindOrganization(orgID)
 		if err != nil || organization == nil {
 			return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeOrganization, nil, err)
@@ -2296,11 +2299,11 @@ func (a *Auth) checkServiceAccountCreds(r *sigauth.Request, accountID *string, f
 
 func (a *Auth) buildAccessTokenForServiceAccount(account model.ServiceAccount, authType string) (string, *model.AppOrgPair, error) {
 	permissions := account.GetPermissionNames()
-	appID := authutils.AllApps
+	appID := rokwireutils.AllApps
 	if account.Application != nil {
 		appID = account.Application.ID
 	}
-	orgID := authutils.AllOrgs
+	orgID := rokwireutils.AllOrgs
 	if account.Organization != nil {
 		orgID = account.Organization.ID
 	}
@@ -2501,11 +2504,16 @@ func (a *Auth) buildRefreshToken() (string, error) {
 // getScopedAccessToken returns a scoped access token with the requested scopes
 func (a *Auth) getScopedAccessToken(claims tokenauth.Claims, serviceID string, scopes []authorization.Scope) (string, error) {
 	aud, scope := a.tokenDataForScopes(scopes)
-	if !authutils.ContainsString(aud, serviceID) {
+	if !rokwireutils.ContainsString(aud, serviceID) {
 		aud = append(aud, serviceID)
 	}
 
-	scopedClaims := a.getStandardClaims(claims.Subject, "", "", "", "", strings.Join(aud, ","), claims.OrgID, claims.AppID, claims.AuthType, claims.ExternalIDs, &claims.ExpiresAt, claims.Anonymous, claims.Authenticated, false, false, claims.Service, false, claims.SessionID)
+	var expiresAt *time.Time
+	if claims.ExpiresAt != nil {
+		expiresAt = &claims.ExpiresAt.Time
+	}
+
+	scopedClaims := a.getStandardClaims(claims.Subject, "", "", "", "", strings.Join(aud, ","), claims.OrgID, claims.AppID, claims.AuthType, claims.ExternalIDs, expiresAt, claims.Anonymous, claims.Authenticated, false, false, claims.Service, false, claims.SessionID)
 	return a.buildAccessToken(scopedClaims, "", scope)
 }
 
@@ -2514,7 +2522,7 @@ func (a *Auth) tokenDataForScopes(scopes []authorization.Scope) ([]string, strin
 	services := []string{}
 	for i, scope := range scopes {
 		scopeStrings[i] = scope.String()
-		if !authutils.ContainsString(services, scope.ServiceID) {
+		if !rokwireutils.ContainsString(services, scope.ServiceID) {
 			services = append(services, scope.ServiceID)
 		}
 	}
@@ -2523,13 +2531,13 @@ func (a *Auth) tokenDataForScopes(scopes []authorization.Scope) ([]string, strin
 }
 
 func (a *Auth) getStandardClaims(sub string, uid string, name string, email string, phone string, aud string, orgID string, appID string,
-	authType string, externalIDs map[string]string, exp *int64, anonymous bool, authenticated bool, admin bool, system bool, service bool, firstParty bool, sessionID string) tokenauth.Claims {
+	authType string, externalIDs map[string]string, exp *time.Time, anonymous bool, authenticated bool, admin bool, system bool, service bool, firstParty bool, sessionID string) tokenauth.Claims {
 	return tokenauth.Claims{
-		StandardClaims: jwt.StandardClaims{
-			Audience:  aud,
+		RegisteredClaims: jwt.RegisteredClaims{
+			Audience:  jwt.ClaimStrings([]string{aud}), // TODO: we originally used a single comma-separated string, but this claim now supports a list of strings. Switching to the list of strings would be a breaking change.
 			Subject:   sub,
-			ExpiresAt: a.getExp(exp),
-			IssuedAt:  time.Now().Unix(),
+			ExpiresAt: jwt.NewNumericDate(a.getExp(exp)),
+			IssuedAt:  jwt.NewNumericDate(time.Now()),
 			Issuer:    a.host,
 		}, OrgID: orgID, AppID: appID, AuthType: authType, UID: uid, Name: name, Email: email, Phone: phone,
 		ExternalIDs: externalIDs, Anonymous: anonymous, Authenticated: authenticated, Admin: admin, System: system,
@@ -2537,19 +2545,19 @@ func (a *Auth) getStandardClaims(sub string, uid string, name string, email stri
 	}
 }
 
-func (a *Auth) getExp(exp *int64) int64 {
+func (a *Auth) getExp(exp *time.Time) time.Time {
 	if exp == nil {
 		defaultTime := time.Now().Add(30 * time.Minute) //TODO: Set up org configs for default token exp
-		return defaultTime.Unix()
+		return defaultTime
 	}
-	expTime := time.Unix(*exp, 0)
+	expTime := exp
 	minTime := time.Now().Add(time.Duration(a.minTokenExp) * time.Minute)
 	maxTime := time.Now().Add(time.Duration(a.maxTokenExp) * time.Minute)
 
 	if expTime.Before(minTime) {
-		return minTime.Unix()
+		return minTime
 	} else if expTime.After(maxTime) {
-		return maxTime.Unix()
+		return maxTime
 	}
 
 	return *exp
@@ -2590,7 +2598,7 @@ func (a *Auth) updateExternalAccountRoles(account *model.Account, newExternalRol
 	newRoles := []model.AccountRole{}
 	//Remove any roles which were not set by an admin and are not in new list
 	for _, role := range account.Roles {
-		if role.AdminSet || authutils.ContainsString(newExternalRoleIDs, role.Role.ID) {
+		if role.AdminSet || rokwireutils.ContainsString(newExternalRoleIDs, role.Role.ID) {
 			newRoles = append(newRoles, role)
 		} else {
 			updated = true
@@ -2637,7 +2645,7 @@ func (a *Auth) updateExternalAccountGroups(account *model.Account, newExternalGr
 	newGroups := []model.AccountGroup{}
 	//Remove any groups which were not set by an admin and are not in new list
 	for _, group := range account.Groups {
-		if group.AdminSet || authutils.ContainsString(newExternalGroupIDs, group.Group.ID) {
+		if group.AdminSet || rokwireutils.ContainsString(newExternalGroupIDs, group.Group.ID) {
 			newGroups = append(newGroups, group)
 		} else {
 			updated = true
@@ -2689,7 +2697,7 @@ func (a *Auth) storeCoreRegs() error {
 	}
 
 	// Setup "auth" registration for token validation
-	authReg := model.ServiceRegistration{Registration: authservice.ServiceReg{ServiceID: authServiceID, Host: a.host, PubKey: a.authPrivKey.PubKey}, CoreHost: a.host,
+	authReg := model.ServiceRegistration{Registration: auth.ServiceReg{ServiceID: authServiceID, Host: a.host, PubKey: a.authPrivKey.PubKey}, CoreHost: a.host,
 		Name: "ROKWIRE Auth Service", Description: "The Auth Service is a subsystem of the Core Building Block that manages authentication and authorization.", FirstParty: true}
 	err = a.storage.SaveServiceReg(&authReg, true)
 	if err != nil {
@@ -2697,7 +2705,7 @@ func (a *Auth) storeCoreRegs() error {
 	}
 
 	// Setup core registration for signature validation
-	coreReg := model.ServiceRegistration{Registration: authservice.ServiceReg{ServiceID: a.serviceID, ServiceAccountID: a.serviceID, Host: a.host, PubKey: a.authPrivKey.PubKey}, CoreHost: a.host,
+	coreReg := model.ServiceRegistration{Registration: auth.ServiceReg{ServiceID: a.serviceID, ServiceAccountID: a.serviceID, Host: a.host, PubKey: a.authPrivKey.PubKey}, CoreHost: a.host,
 		Name: "ROKWIRE Core Building Block", Description: "The Core Building Block manages user, auth, and organization data for the ROKWIRE platform.", FirstParty: true}
 	err = a.storage.SaveServiceReg(&coreReg, true)
 	if err != nil {
@@ -2939,13 +2947,13 @@ func (a *Auth) deleteDeletedMemberships() {
 // LocalServiceRegLoaderImpl provides a local implementation for ServiceRegLoader
 type LocalServiceRegLoaderImpl struct {
 	storage Storage
-	*authservice.ServiceRegSubscriptions
+	*auth.ServiceRegSubscriptions
 }
 
 // LoadServices implements ServiceRegLoader interface
-func (l *LocalServiceRegLoaderImpl) LoadServices() ([]authservice.ServiceReg, error) {
+func (l *LocalServiceRegLoaderImpl) LoadServices() ([]auth.ServiceReg, error) {
 	regs := l.storage.FindServiceRegs(l.GetSubscribedServices())
-	authRegs := make([]authservice.ServiceReg, len(regs))
+	authRegs := make([]auth.ServiceReg, len(regs))
 	for i, serviceReg := range regs {
 		reg := serviceReg.Registration
 		reg.PubKey.Decode()
@@ -2957,32 +2965,32 @@ func (l *LocalServiceRegLoaderImpl) LoadServices() ([]authservice.ServiceReg, er
 
 // NewLocalServiceRegLoader creates and configures a new LocalServiceRegLoaderImpl instance
 func NewLocalServiceRegLoader(storage Storage) *LocalServiceRegLoaderImpl {
-	subscriptions := authservice.NewServiceRegSubscriptions([]string{allServices})
+	subscriptions := auth.NewServiceRegSubscriptions([]string{allServices})
 	return &LocalServiceRegLoaderImpl{storage: storage, ServiceRegSubscriptions: subscriptions}
 }
 
-// LocalServiceAccountLoaderImpl provides a local implementation for authservice.ServiceAccountLoader
+// LocalServiceAccountLoaderImpl provides a local implementation for auth.ServiceAccountLoader
 type LocalServiceAccountLoaderImpl struct {
 	auth Auth
 }
 
 // LoadAccessToken gets an access token for appID, orgID if the implementing service is granted access
-func (l *LocalServiceAccountLoaderImpl) LoadAccessToken(appID string, orgID string) (*authservice.AccessToken, error) {
-	account, err := l.auth.storage.FindServiceAccount(nil, l.auth.serviceID, authutils.AllApps, authutils.AllOrgs)
+func (l *LocalServiceAccountLoaderImpl) LoadAccessToken(appID string, orgID string) (*auth.AccessToken, error) {
+	account, err := l.auth.storage.FindServiceAccount(nil, l.auth.serviceID, rokwireutils.AllApps, rokwireutils.AllOrgs)
 	if err != nil || account == nil {
 		return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeServiceAccount, logutils.StringArgs(l.auth.serviceID), err)
 	}
 	token, _, err := l.auth.buildAccessTokenForServiceAccount(*account, ServiceAuthTypeCore)
-	return &authservice.AccessToken{Token: token, TokenType: model.TokenTypeBearer}, err
+	return &auth.AccessToken{Token: token, TokenType: model.TokenTypeBearer}, err
 }
 
 // LoadAccessTokens gets an access token for each app org pair the implementing service is granted access
-func (l *LocalServiceAccountLoaderImpl) LoadAccessTokens() (map[authservice.AppOrgPair]authservice.AccessToken, error) {
-	token, err := l.LoadAccessToken(authutils.AllApps, authutils.AllOrgs)
+func (l *LocalServiceAccountLoaderImpl) LoadAccessTokens() (map[auth.AppOrgPair]auth.AccessToken, error) {
+	token, err := l.LoadAccessToken(rokwireutils.AllApps, rokwireutils.AllOrgs)
 	if err != nil || token == nil {
 		return nil, errors.WrapErrorAction(logutils.ActionGet, logutils.TypeToken, nil, err)
 	}
-	tokens := map[authservice.AppOrgPair]authservice.AccessToken{{AppID: authutils.AllApps, OrgID: authutils.AllOrgs}: *token}
+	tokens := map[auth.AppOrgPair]auth.AccessToken{{AppID: rokwireutils.AllApps, OrgID: rokwireutils.AllOrgs}: *token}
 	return tokens, nil
 }