+++ driven/identitybb/adapter.go
@@ -21,17 +21,17 @@ import (
 	"net/http"
 	"net/url"
 
-	"github.com/rokwire/core-auth-library-go/v3/authservice"
+	"github.com/rokwire/rokwire-building-block-sdk-go/services/core/auth"
 
-	"github.com/rokwire/core-auth-library-go/v3/authutils"
-	"github.com/rokwire/logging-library-go/v2/errors"
-	"github.com/rokwire/logging-library-go/v2/logs"
-	"github.com/rokwire/logging-library-go/v2/logutils"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/errors"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/logging/logs"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/logging/logutils"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/rokwireutils"
 )
 
 // Adapter implements the IdentityBuildingBlock interface
 type Adapter struct {
-	serviceAccountManager *authservice.ServiceAccountManager
+	serviceAccountManager *auth.ServiceAccountManager
 }
 
 // GetUserProfile gets user profile info for the provided user credentials
@@ -57,7 +57,7 @@ func (a *Adapter) GetUserProfile(baseURL string, externalUser model.ExternalSyst
 
 	req.Header.Set("External-Authorization", externalAccessToken)
 
-	resp, err := a.serviceAccountManager.MakeRequest(req, authutils.AllApps, authutils.AllOrgs)
+	resp, err := a.serviceAccountManager.MakeRequest(req, rokwireutils.AllApps, rokwireutils.AllOrgs)
 	if err != nil {
 		return nil, errors.WrapErrorAction(logutils.ActionSend, logutils.TypeRequest, nil, err)
 	}
@@ -86,6 +86,6 @@ func (a *Adapter) GetUserProfile(baseURL string, externalUser model.ExternalSyst
 }
 
 // NewIdentityBBAdapter creates a new identity building block adapter instance
-func NewIdentityBBAdapter(serviceAccountManager *authservice.ServiceAccountManager) *Adapter {
+func NewIdentityBBAdapter(serviceAccountManager *auth.ServiceAccountManager) *Adapter {
 	return &Adapter{serviceAccountManager: serviceAccountManager}
 }