+++ utils/utils.go
@@ -26,10 +26,10 @@ import (
 	"strings"
 	"time"
 
-	"github.com/rokwire/logging-library-go/v2/logs"
-	"github.com/rokwire/logging-library-go/v2/logutils"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/logging/logs"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/logging/logutils"
 
-	"github.com/rokwire/logging-library-go/v2/errors"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/errors"
 )
 
 const (